+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.4-SNAPSHOT
+
+- (start) or (start schema storage)
+  if exists `storage`, ignores `schema`.
+
 # 0.3.54 (2025-08-19)
 
 - {:git/tag :git/sha} - use `git show-ref --abbrev=7 --tags`